requirements.txt
discord.py==1.3.4
PyNaCl==1.3.0
dnspython==1.16.0
requests
bs4
async-timeout==3.0.1